Add files via upload
You are a data scientist working for a hedge fund; it's your job to determine any suspicious stock activity. 
In this lab you will extract stock data using a Python library. We will use the yfinance library, it allows us to extract data for stocks returning data in a pandas dataframe. 
You will use the lab to extract.

- Using yfinance to Extract Stock Info
- Using yfinance to Extract Historical Share Price Data
- Using yfinance to Extract Historical Dividends Data
- Exercise